Handling  last  state
